+++ app/scripts/fhir/R4/sdc-import.js
@@ -621,53 +621,6 @@ function addSDCImportFns(ns) {
   };
 
 
-  /**
-   * Parse questionnaire item for display control
-   *
-   * @param lfItem {object} - LForms item object to assign display control
-   * @param qItem {object} - Questionnaire item object
-   * @private
-   */
-  self._processDisplayControl = function (lfItem, qItem) {
-    var itemControlType = LForms.Util.findObjectInArray(qItem.extension, 'url', self.fhirExtUrlItemControl);
-
-    if(itemControlType) {
-      var displayControl = {};
-      switch (itemControlType.valueCodeableConcept.coding[0].code) {
-        case 'Lookup':
-          // TODO -
-          // Implies externallyDefined, but the URL is not saved in fhir resource.
-          // Perhaps it could be save in itemControlType.valueCodableConcept.text ...
-          // lfItem.externallyDefined = itemControlType.valueCodableConcept.text;
-          break;
-        case 'Combo-box':
-          displayControl.answerLayout = {type: 'COMBO_BOX'};
-          break;
-        case 'Checkbox':
-        case 'Radio':
-          displayControl.answerLayout = {type: 'RADIO_CHECKBOX'};
-          break;
-        case 'Table':
-          if(lfItem.dataType === 'SECTION') {
-            displayControl.questionLayout = "horizontal";
-          }
-          break;
-        case 'Matrix':
-          if(lfItem.dataType === 'SECTION') {
-            displayControl.questionLayout = "matrix";
-          }
-          break;
-        default:
-          displayControl = null;
-      }
-
-      if(displayControl && !jQuery.isEmptyObject(displayControl)) {
-        lfItem.displayControl = displayControl;
-      }
-    }
-  };
-
-
   // Quesitonnaire Response Import
   self._mergeQR = {
 
@@ -689,19 +642,6 @@ function addSDCImportFns(ns) {
 
 
     /**
-     * Get the item code from a link id
-     * @param linkId a link id
-     * @returns {*}
-     * @private
-     */
-    _getItemCodeFromLinkId: function(linkId) {
-      var parts = linkId.split("/");
-      var itemCode = parts[parts.length -1];
-      return itemCode;
-    },
-
-
-    /**
      * Get structural info of a QuestionnaireResponse by going though each level of items
      * @param parentQRItemInfo the structural info of a parent item
      * @param parentItem a parent item in a QuestionnaireResponse object
@@ -715,16 +655,16 @@ function addSDCImportFns(ns) {
       if (parentQRItem && parentQRItem.item) {
         for (var i=0, iLen=parentQRItem.item.length; i<iLen; i++) {
           var item = parentQRItem.item[i];
-          var itemCode = this._getItemCodeFromLinkId(item.linkId);
+          var linkId = item.linkId; //code is not necessary included in linkId
           // first item that has the same code, either repeating or non-repeating
-          if (!repeatingItemProcessed[itemCode]) {
-            var repeatingInfo = this._findTotalRepeatingNum(itemCode, parentQRItem);
+          if (!repeatingItemProcessed[linkId]) {
+            var repeatingInfo = this._findTotalRepeatingNum(linkId, parentQRItem);
 
             // create structure info for the item
             var repeatingItems = repeatingInfo.repeatingItems;
             for (var j=0, jLen=repeatingItems.length; j<jLen; j++) {
               var qrItemInfo = {
-                code: itemCode,
+                linkId: linkId,
                 item: repeatingItems[j],
                 index: j,
                 total: repeatingInfo.total
@@ -733,7 +673,7 @@ function addSDCImportFns(ns) {
               this._checkQRItems(qrItemInfo, repeatingItems[j]);
               qrItemsInfo.push(qrItemInfo);
             }
-            repeatingItemProcessed[itemCode] = true;
+            repeatingItemProcessed[linkId] = true;
           }
         }
         parentQRItemInfo.qrItemsInfo = qrItemsInfo;
@@ -743,19 +683,18 @@ function addSDCImportFns(ns) {
 
     /**
      * Find the number of the repeating items that have the same code
-     * @param code an item code
+     * @param linkId an item's linkId
      * @param parentQRItem a parent item in a QuestionnaireResponse object
      * @returns a structural info object for a repeating item
      * @private
      */
-    _findTotalRepeatingNum : function(code, parentQRItem) {
+    _findTotalRepeatingNum : function(linkId, parentQRItem) {
 
       var total = 0;
       var repeatingItems = [];
       for (var i=0, iLen=parentQRItem.item.length; i<iLen; i++) {
         var item = parentQRItem.item[i];
-        var itemCode = this._getItemCodeFromLinkId(item.linkId);
-        if (itemCode === code) {
+        if (linkId === item.linkId) {
           repeatingItems.push(item);
           if (Array.isArray(item.answer)) {
             total += item.answer.length; // answers for repeating questions and repeating answers
@@ -776,16 +715,16 @@ function addSDCImportFns(ns) {
     /**
      * Add repeating items into LForms definition data object
      * @param parentItem a parent item
-     * @param itemCode code of a repeating item
+     * @param linkId linkId of a repeating item
      * @param total total number of the repeating item with the same code
      * @private
      */
-    _addRepeatingItems : function(parentItem, itemCode, total) {
+    _addRepeatingItems : function(parentItem, linkId, total) {
       // find the first (and the only one) item
       var item = null;
       if (parentItem.items) {
         for(var i=0, iLen=parentItem.items.length; i<iLen; i++) {
-          if (itemCode === parentItem.items[i].questionCode) {
+          if (linkId === parentItem.items[i].linkId) {
             item = parentItem.items[i];
             break;
           }
@@ -805,17 +744,17 @@ function addSDCImportFns(ns) {
     /**
      * Find a matching repeating item by item code and the index in the items array
      * @param parentItem a parent item
-     * @param itemCode code of a repeating (or non-repeating) item
+     * @param linkId linkId of a repeating (or non-repeating) item
      * @param index index of the item in the sub item array of the parent item
      * @returns {{}} a matching item
      * @private
      */
-    _findTheMatchingItemByCodeAndIndex : function(parentItem, itemCode, index) {
+    _findTheMatchingItemByLinkIdAndIndex : function(parentItem, linkId, index) {
       var item = null;
       var idx = 0;
       if (parentItem.items) {
         for(var i=0, iLen=parentItem.items.length; i<iLen; i++) {
-          if (itemCode === parentItem.items[i].questionCode) {
+          if (linkId === parentItem.items[i].linkId) {
             if (idx === index) {
               item = parentItem.items[i];
               break;
@@ -834,15 +773,15 @@ function addSDCImportFns(ns) {
      * Find a matching repeating item by item code alone
      * When used on the LForms definition data object, there is no repeating items yet.
      * @param parentItem a parent item
-     * @param itemCode code of an item
+     * @param linkId linkId of an item
      * @returns {{}} a matching item
      * @private
      */
-    _findTheMatchingItemByCode : function(parentItem, itemCode) {
+    _findTheMatchingItemByLinkId : function(parentItem, linkId) {
       var item = null;
       if (parentItem.items) {
         for(var i=0, iLen=parentItem.items.length; i<iLen; i++) {
-          if (itemCode === parentItem.items[i].questionCode) {
+          if (linkId === parentItem.items[i].linkId) {
             item = parentItem.items[i];
             break;
           }
@@ -854,14 +793,14 @@ function addSDCImportFns(ns) {
 
     /**
      * Set value and units on a LForms item
-     * @param code an item code
+     * @param linkId an item's linkId
      * @param answer value for the item
      * @param item a LForms item
      * @private
      */
-    _setupItemValueAndUnit : function(code, answer, item) {
+    _setupItemValueAndUnit : function(linkId, answer, item) {
 
-      if (item && code === item.questionCode && (item.dataType !== 'SECTION' && item.dataType !== 'TITLE')) {
+      if (item && linkId === item.linkId && (item.dataType !== 'SECTION' && item.dataType !== 'TITLE')) {
         var dataType = item.dataType;
 
         // any one has a unit must be a numerical type, let use REAL for now.